refactor: rename index.tsx to PascalCase.tsx, add barrel index.tsx per component
index.tsx files now re-export from the named component file, matching
the Tasklist convention for component folder structure.
