+++ webapp/client/apps/orchestration-cluster-webapp/src/operate/shared/DiagramShell/index.tsx
@@ -6,64 +6,4 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import React from 'react';
-import {Loading} from '@carbon/react';
-import {Container, LoadingContainer, EmptyMessage, ErrorMessage} from './styled';
-import {EmptyMessage as BaseEmptyMessage} from '../EmptyMessage';
-
-type DefaultProps = {
-	children: React.ReactNode;
-	status: 'error' | 'loading' | 'content' | 'forbidden';
-};
-
-type WithEmptyMessageProps = {
-	children: React.ReactNode;
-	status: 'error' | 'empty' | 'loading' | 'content';
-	emptyMessage: React.ComponentProps<typeof BaseEmptyMessage>;
-	messagePosition?: 'top' | 'center';
-};
-
-const DiagramShell: React.FC<DefaultProps | WithEmptyMessageProps> = ({children, status, ...props}) => (
-	<Container data-testid="diagram-body" tabIndex={0}>
-		{(() => {
-			if (status === 'content') {
-				return children;
-			}
-
-			if (status === 'loading') {
-				return (
-					<>
-						<LoadingContainer>
-							<Loading data-testid="diagram-spinner" withOverlay={false} />
-						</LoadingContainer>
-						{children}
-					</>
-				);
-			}
-
-			if (status === 'empty' && 'emptyMessage' in props) {
-				return <EmptyMessage $position={props.messagePosition} {...props.emptyMessage} />;
-			}
-
-			if (status === 'forbidden') {
-				const position = 'messagePosition' in props ? props.messagePosition : 'top';
-				return (
-					<ErrorMessage
-						$position={position}
-						message="Missing permissions to view the Definition"
-						additionalInfo="Please contact your organization owner or admin to give you the necessary permissions to read this definition"
-					/>
-				);
-			}
-
-			if (status === 'error') {
-				const position = 'messagePosition' in props ? props.messagePosition : 'top';
-				return <ErrorMessage $position={position} />;
-			}
-
-			return null;
-		})()}
-	</Container>
-);
-
-export {DiagramShell};
+export {DiagramShell} from "./DiagramShell";
