+++ webapp/client/apps/orchestration-cluster-webapp/src/operate/shared/ErrorMessage/index.tsx
@@ -6,21 +6,4 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import React from 'react';
-import {EmptyMessage} from '../EmptyMessage';
-
-type Props = {
-	message?: string;
-	additionalInfo?: string;
-};
-
-const DEFAULT_ERROR = {
-	message: 'Data could not be fetched',
-	additionalInfo: 'Refresh the page to try again',
-};
-
-const ErrorMessage: React.FC<Props> = (props) => {
-	return <EmptyMessage {...DEFAULT_ERROR} {...props} />;
-};
-
-export {ErrorMessage};
+export {ErrorMessage} from "./ErrorMessage";
